Enable Linux System for Windows 10
With the update to Windows 10 Fall Creators edition WSL is out of beta
so I think it should be safe to enable it (which is helpful because
doing so requires a reboot). I'm leaving it off the Cygwin distro
because it seems you have already made you bed there. I'm also leaving
it off on the SSH flavor as well (for no good reason really).